+++ client/src/components/Clientcard/index.jsx
@@ -1,4 +1,4 @@
-import clientcard from './Clientcard.css';
+import './Clientcard.css';
