+++ client/src/components/Footer/Footer.jsx
@@ -1,4 +1,4 @@
-import footer from './Footer.css';
+import './Footer.css';
 
 
 export default function Footer() {
