+++ client/src/components/Header/index.jsx
@@ -1,13 +1,13 @@
-import header from './Header.css';
+import './Header.css';
 
 export default function Header() {
     return (
-        <div>
+        <header>
             <h1>Excalibur</h1>
             <p>
                 Welcome to Excalibur. Join our Round Table where freelancers can offer complete transparency in the cost of enhancing a companies technilogical footprint and ability.
 
             </p>
-        </div>
+        </header>
     )
 }
